+++ cedar-policy/CHANGELOG.md
@@ -22,9 +22,8 @@ Cedar Language Version: TBD
   format. The new functions are deprecated and placed behind the `deprecated-schema-compat` feature. (#1600)
 - `Expression::new_duration`, `Expression::new_datetime`, `RestrictedExpression::new_duration`,
    and `RestrictedExpression::new_datetime` (#1614)
-- Added a function to be able to split a policy set parsed from a large file into its component static policies
-  and templates. The relevant functions are `policy_set_text_to_parts` which can be used from wasm, and 
-  `PolicySet::to_string_representations` in `api.rs`, which returns a `StringifiedPolicySet`.
+- Added a function to be able to split a policy set parsed from a single string into its component static 
+  policies and templates. The relevant function is `policy_set_text_to_parts` in the `ffi` module.
 
 ### Changed